Base code commit
The base code for the importer to be uploaded.

This version is not fully functional, it has been tested against a
reasonable test case however additional testing still required.

Certain modules have been disabled and will be added in the future.

Additionally a sample config file has NOT been included and will be
added in the next push.